+++ Flippy/FLToolbarNode.m
@@ -12,6 +12,8 @@
 
 #import "FLToolbarNode.h"
 
+#include <tgmath.h>
+
 #import "FLTextureStore.h"
 
 static UIColor *FLToolbarColorBackground;
@@ -404,10 +406,14 @@ static UIColor *FLToolbarColorButtonHighlighted;
 
 CGSize FL_rotatedSizeBounds(CGSize size, CGFloat theta)
 {
-  CGFloat widthRotatedWidth = size.width * cosf(theta);
-  CGFloat widthRotatedHeight = size.width * sinf(theta);
-  CGFloat heightRotatedWidth = size.height * sinf(theta);
-  CGFloat heightRotatedHeight = size.height * cosf(theta);
+  // note: These casts back to CGFloat are supposed to be unnecessary because of
+  // the type-generic cos() macro defined in tgmath.h.  But I get loss-of-precision
+  // warnings when compiling for 32-bit simulator without the casts; I'm not sure who
+  // is to blame.
+  CGFloat widthRotatedWidth = size.width * (CGFloat)cos(theta);
+  CGFloat widthRotatedHeight = size.width * (CGFloat)sin(theta);
+  CGFloat heightRotatedWidth = size.height * (CGFloat)sin(theta);
+  CGFloat heightRotatedHeight = size.height * (CGFloat)cos(theta);
   return CGSizeMake(widthRotatedWidth + heightRotatedWidth,
                     widthRotatedHeight + heightRotatedHeight);
 }
